Adjust logging
Most things were set on DEBUG, with no possibility to lower them.

Changed a gauge message that would log ERROR on a simple lack of update
for a topic to DEBUG.  Changed general logging to an env variable
(LOG_LEVEL) to allow override at runtime.  Reduced logging for
com.amazonaws, which was giving lots of logs when sending to cloudwatch.